SearchRequest renamed and working
