+++ src/entities/libraries/distributedLibrary.ts
@@ -1,21 +1,21 @@
-import {IMoney} from "../../valueItems/money/IMoney";
-import {IBorrower} from "../people/IBorrower";
-import {IThing} from "../things/IThing";
-import {ThingStatus} from "../../valueItems/thingStatus";
-import {ILoan} from "../loans/ILoan";
-import {Loan} from "../loans/loan"
-import {LoanStatus} from "../../valueItems/loanStatus";
-import {BorrowerNotInGoodStandingError, InvalidThingStatusToBorrowError} from "../../valueItems/exceptions";
-import {ThingTitle} from "../../valueItems/thingTitle";
+import {IMoney} from "../../valueItems";
+import {IBorrower} from "../people";
+import {IThing} from "../things";
+import {ThingStatus} from "../../valueItems";
+import {ILoan} from "../loans";
+import {Loan} from "../loans"
+import {LoanStatus} from "../../valueItems";
+import {BorrowerNotInGoodStandingError, InvalidThingStatusToBorrowError} from "../../valueItems";
+import {ThingTitle} from "../../valueItems";
 import {BaseLibrary} from "./baseLibrary";
-import {IWaitingListFactory} from "../../factories/IWaitingListFactory";
-import {Person} from "../people/person";
-import {DueDate} from "../../valueItems/dueDate";
-import {IFeeSchedule} from "../../factories/IFeeSchedule";
-import {ILender} from "../lenders/ILender";
-import {TimeInterval} from "../../valueItems/timeInterval";
-import {QuadraticBiddingStrategy} from "../../services/bidding/quadraticBiddingStrategy";
-import {MoneyFactory} from "../../factories/moneyFactory";
+import {IWaitingListFactory} from "../../factories";
+import {Person} from "../people";
+import {DueDate} from "../../valueItems";
+import {IFeeSchedule} from "../../factories";
+import {ILender} from "../lenders";
+import {TimeInterval} from "../../valueItems";
+import {QuadraticBiddingStrategy} from "../../services";
+import {MoneyFactory} from "../../factories";
 
 export class DistributedLibrary extends BaseLibrary{
     private readonly _lenders: ILender[]
