+++ src/factories/feeSchedules/index.ts
@@ -1,0 +1,2 @@
+export * from "./IFeeSchedule"
+export * from "./simpleTimeBasedFeeSchedule"
