+++ src/factories/index.ts
@@ -1,5 +1,4 @@
-export * from "./IFeeSchedule"
+export * from "./feeSchedules"
 export * from "./IWaitingListFactory"
 export * from "./moneyFactory"
-export * from "./simpleTimeBasedFeeSchedule"
 export * from "./waitingListFactory"
