+++ src/services/titleSearch/titleSearchService.ts
@@ -1,7 +1,7 @@
 import {ITitleSearchService} from "./ITitleSearchService"
 import {Person} from "../../entities"
 import {ILibraryRepository} from "../../repositories"
-import {TitleSearchRequest, ThingTitle} from "../../valueItems/"
+import {ThingTitle, TitleSearchRequest} from "../../valueItems/"
 
 export class TitleSearchService implements ITitleSearchService {
     private readonly libraryRepository: ILibraryRepository
@@ -10,16 +10,27 @@ export class TitleSearchService implements ITitleSearchService {
         this.libraryRepository = libraryRepository
     }
 
+    private matches(searchRequest: TitleSearchRequest, title: ThingTitle): boolean {
+        if(!searchRequest.searchText){
+            return true;
+        }
+        return title.name.includes(searchRequest.searchText) ||
+            title.isbn == searchRequest.searchText ||
+            title.upc == searchRequest.searchText;
+    }
+
     * find(person: Person, searchRequest: TitleSearchRequest): Iterable<ThingTitle> {
         const libraries = this.libraryRepository.getLibrariesPersonCanUse(person);
 
         const exported = []
         for(const library of libraries) {
-            for(const item of library.availableTitles){
-                const id = item.hash
+            for(const title of library.availableTitles){
+                const id = title.hash
                 if(exported.indexOf(id) < 0){
-                    exported.push(id)
-                    yield item
+                    if(this.matches(searchRequest, title)) {
+                        exported.push(id)
+                        yield title
+                    }
                 }
             }
         }
